web/nginx_serpent/deploy/flask_app/requirements.txt
docker==7.1.0
Flask==3.0.3
gunicorn==21.2.0
gevent==23.9.1
python-dotenv==1.0.0
Pillow==9.5.0
psycopg2-binary==2.9.9  # PostgreSQL adapter
prometheus-client==0.19.0  # For exposing Prometheus metrics
psutil==5.9.8  # For system resource monitoring
captcha==0.4
